Revert "se leen datos de inicializacion por sd tiene bugs"
This reverts commit 1b140b088aeaeb070f1bd31e648cd8247bd28f2c.
